Field_Extraction/requirements.txt
streamlit
azure-ai-documentintelligence
openai
python-dotenv
pydantic